20161121 update

+++ Info360/src/com/infotrends/servlet/Delete_PersonInfoServlet.java
@@ -1,6 +1,7 @@
 package com.infotrends.servlet;
 
 import java.io.IOException;
+import java.util.List;
 
 import javax.ws.rs.FormParam;
 import javax.ws.rs.POST;
@@ -10,9 +11,11 @@ import javax.ws.rs.core.Response;
 
 import org.json.JSONObject;
 
+import com.infotrends.bean.CFG_group_person;
 import com.infotrends.bean.CFG_person;
 import com.infotrends.service.MaintainService;
 import com.infotrends.util.IsError;
+import com.infotrends.util.Variable;
 
 
 @Path("/Delete_PersonInfo")
@@ -28,23 +31,33 @@ public class Delete_PersonInfoServlet {
 	
 	@POST
 	@Produces("application/json")
-    public Response postFromPath(@FormParam("ACCOUNT") String account,
-			@FormParam("PASSWORD") String password
+    public Response postFromPath(@FormParam("dbid") int dbid,
+			@FormParam("password") String password
 			) throws IOException {
 		
 
 		JSONObject jsonObject = new JSONObject();
 		CFG_person cfg_person = new CFG_person();
-		jsonObject.put("Status", "POST Path");
-		cfg_person.setAccount(account);
-		cfg_person.setPassword(password);
-	
+		jsonObject.put("Status", Variable.POST_STATUS);
+		//cfg_person.setAccount(account);
+		//cfg_person.setPassword(password);
+		cfg_person.setDbid(dbid);
 		
-		int deletecount=0;
 		try{
 			MaintainService maintainService = new MaintainService();
-			deletecount = maintainService.delete_PersonInfo(cfg_person);
-			jsonObject.put("deletecount", deletecount);
+			List<CFG_person> cfg_personlist = maintainService.query_Person_DBID(cfg_person);
+			if(password.trim().equals(cfg_personlist.get(0).getPassword().trim())){
+				cfg_person.setPassword(password);
+				int deletepersoncount = maintainService.delete_PersonInfo(cfg_person);
+				jsonObject.put("delete_personcount", deletepersoncount);
+				CFG_group_person cfg_group_person = new CFG_group_person();
+				cfg_group_person.setPerson_dbid(cfg_personlist.get(0).getDbid());
+				int deletegrouppersoncount =maintainService.delete_Group_PersonInfo(cfg_group_person);
+				jsonObject.put("delete_group_personcount", deletegrouppersoncount);
+			}else{
+				jsonObject.put("error", "wrong password");
+			}
+			
 			
 		} catch (Exception e) {
 			if(IsError.GET_EXCEPTION != null)
